عدم استفاده از ماژول عمومی چک کردن موجودی در انتقالی انبار و چک کردن عمیق تر در حساب ها در حذف

+++ Prg_UI/Wins/WinMenus/Taarif/DETA_HES_SHEET.xaml.cs
@@ -399,9 +399,18 @@ namespace Wins.WinMenus.Taarif
 
                                             try
                                             {
-                                                IsDeletedSomething = true;
-
-                                                dbms.DoExecuteSQL($@"DELETE FROM dbo.DETA_HES WHERE ID = {_id}");
+                                                var tafChildCount = dbms.DoGetDataSQL<int>($@"SELECT COUNT(*) FROM dbo.TDETA_HES WHERE N_KOL = {N_KOL} AND NUMBER = {_NUMBER}").FirstOrDefault();
+                                                var gerdeshCount = dbms.DoGetDataSQL<int>($@"SELECT COUNT(*) FROM dbo.DEED_DTL WHERE HES_K = {N_KOL} AND HES_M = {_NUMBER}").FirstOrDefault();
+                                                if (tafChildCount > 0 || gerdeshCount > 0)
+                                                {
+                                                    e.Handled = true;
+                                                    ErrosMessages.Add(new MsgModel { MessageText_U = $"این حساب دارای گردش است و نمیتوان آنرا حذف کرد!" });
+                                                }
+                                                else
+                                                {
+                                                    dbms.DoExecuteSQL($@"DELETE FROM dbo.DETA_HES WHERE ID = {_id}");
+                                                    IsDeletedSomething = true;
+                                                }
                                             }
                                             catch (SqlException ex)
                                             {
